Move tabBarView out of teachWordView
